refactor: extract Calendar/Plan/Log tabs to views/ + render tests
Moves the three remaining inline tab blocks out of the erg-dashboard.jsx
shell into views/CalendarView.jsx, PlanView.jsx, and LogView.jsx (logic
moved verbatim, props threaded from App). Removes the now-unused imports
(WorkoutItem, LogEntry, LogSessionForm, SRPE_GUIDE, MICROCYCLE, EVENT_LADDER,
PHASE_CONTEXT, and the utils/schedule roster helpers) from the monolith,
which drops from 963 to 516 lines. Adds a render test per new view; each
file is counted by the explicit coverage scope (no exclude entry needed).

Closes #78
